Add randomization to urls required to fetch all nearby pages
- If #required nearby pages < #available nearby pages, then randomize the urls in order to avoid biasing of results in one direction. Since the small circles are stacked in a hexagonal manner (and it stacks from south to north), not randomizing urls will result in cluttering of results in south with no result on north, in case # required pages < # available pages

- moved `fetch_url` to utils.py

- removed filtering based on distance on frontend side for nearby point rendering
