+++ backend/utils.py
@@ -1,4 +1,6 @@
 import math
+import httpx
+
 def generate_circle_centers(center_lat, center_lon, radius_km, small_radius_km=10):
     """
         Generate a list of centers of small circles (radius=10km) needed to cover a larger circle.
@@ -41,3 +43,27 @@ def generate_circle_centers(center_lat, center_lon, radius_km, small_radius_km=1
                 results.append((lat, lon))
     
     return results
+
+
+
+async def fetch_url(client: httpx.AsyncClient, url: str):
+    """
+        Fetch a URL asynchronously using httpx and return the response status and data.
+        This function is asynchrounously used to fetch multiple URLs in parallel when search radius > 10km.
+        Input:
+        - client: httpx.AsyncClient instance
+        - url: URL to fetch
+        Output:
+        - A dictionary with the URL, status code, and data if available.
+            - Data includes the JSON format of wiki geosearch response.
+        If an error occurs, return a dictionary with the URL and the error message.
+    """
+    try:
+        response = await client.get(url, timeout=10.0)
+        return {
+            "url": url,
+            "status": response.status_code,
+            "data": response.json() if response.status_code == 200 else None,
+        }
+    except Exception as e:
+        return {"url": url, "error": str(e)}
