+++ main.py
@@ -8,7 +8,8 @@ import geopy.distance
 from cachetools import TTLCache
 import os
 from dotenv import load_dotenv
-from backend.utils import generate_circle_centers
+from random import shuffle
+from backend.utils import generate_circle_centers, fetch_url
 
 load_dotenv()
 
@@ -175,16 +176,7 @@ def get_geodistance(payload: Geodistance):
     )
 
 
-async def fetch_url(client: httpx.AsyncClient, url: str):
-    try:
-        response = await client.get(url, timeout=10.0)
-        return {
-            "url": url,
-            "status": response.status_code,
-            "data": response.json() if response.status_code == 200 else None,
-        }
-    except Exception as e:
-        return {"url": url, "error": str(e)}
+
     
 @app.post("/wiki/nearby")
 async def get_nearby_wiki_pages(payload: NearbyWikiPage):
@@ -217,7 +209,9 @@ async def get_nearby_wiki_pages(payload: NearbyWikiPage):
     radius = payload.radius
     limit = payload.limit
 
-    if radius <= 10000:
+    wiki_geosearch_radius_limit_meters = 10000 # Wikipedia API limit for geosearch radius in meters
+
+    if radius <= wiki_geosearch_radius_limit_meters:
         url = ("https://en.wikipedia.org/w/api.php"+"?action=query"
                 "&list=geosearch"
                 f"&gscoord={lat_center}|{lon_center}"
@@ -248,11 +242,14 @@ async def get_nearby_wiki_pages(payload: NearbyWikiPage):
                 content={"error": str(e)},
                 status_code=500
             )
-    elif radius > 10000:
+        
+    elif radius > wiki_geosearch_radius_limit_meters:
+        print(radius)
         small_circle_centers = generate_circle_centers(lat_center, lon_center, radius / 1000, small_radius_km=10)
         all_pages = []
         base_url = "https://en.wikipedia.org/w/api.php?action=query&list=geosearch&gscoord={lat}|{lon}&gsradius={small_radius_km}&gslimit={page_limit}&format=json"
-        urls = [base_url.format(lat=center[0], lon=center[1], small_radius_km=10*1000, page_limit=100) for center in small_circle_centers]
+        urls = [base_url.format(lat=center[0], lon=center[1], small_radius_km=wiki_geosearch_radius_limit_meters, page_limit=100) for center in small_circle_centers]
+        shuffle(urls) # If # available pages > # requested pages by user, randomize the results to avoid clustering around a single direction.
 
         print("URL Counts:", len(urls))
         try:
@@ -263,17 +260,24 @@ async def get_nearby_wiki_pages(payload: NearbyWikiPage):
             # print(results)
             for result in results:
                 for unit in result.get("data", {}).get("query", {}).get("geosearch", []):
+                    if len(all_pages) >= limit:
+                        break
                     lat, lon = unit.get("lat"), unit.get("lon")
                     if lat is not None and lon is not None:
                         dist = int(geopy.distance.distance(
                                 (lat_center, lon_center), (lat, lon)
                             ).m)
-                        print(dist)
+                        # print(dist)
                     else: 
                         dist = None
 
+                    if (not dist) or (dist and dist > radius):
+                        continue
+
                     unit_with_dist = {**unit, "dist": dist}
                     all_pages.append(unit_with_dist)
+
+            # print(all_pages)
 
             return JSONResponse(
                 content={
